+++ R/print.meta_rn.R
@@ -4,7 +4,6 @@
 #' Printing method for meta_rn object.
 #'
 #' @param x \code{meta_rn} object simulated with \code{run_meta}.
-#' @param timestep Numeric with timestep to print.
 #' @param digits Numeric of decimal places (passed on to \code{round}).
 #' @param ... Not used.
 #'
@@ -21,16 +20,10 @@
 #' @rdname print.meta_rn
 #'
 #' @export
-print.meta_rn <- function(x, digits = 2, timestep = x$max_i, ...) {
-
-  # get timestep to be printed
-  timestep_temp <- timestep
+print.meta_rn <- function(x, digits = 3, ...) {
 
   # calc biomass values
-  biomass <- lapply(x$seafloor, function(i) {
-
-    # get selected timestep
-    seafloor_temp <- subset(i, timestep == timestep_temp)
+  biomass <- lapply(x$seafloor, function(seafloor_temp) {
 
     # calculate total values
     c(bg = round(mean(seafloor_temp$bg_biomass, na.rm = TRUE), digits = digits),
@@ -41,22 +34,18 @@ print.meta_rn <- function(x, digits = 2, timestep = x$max_i, ...) {
   })
 
   # get number of reefs
-  no_reefs <- vapply(x$seafloor, function(i) {
-
-    # only get first timestep because reefs are identical
-    seafloor_temp <- subset(i, timestep == 0)
+  no_reefs <- vapply(x$seafloor, function(seafloor_temp) {
 
     # get number of rows in which reef = 1
-    nrow(seafloor_temp[seafloor_temp$reef == 1, ])}, FUN.VALUE = numeric(1))
+    nrow(seafloor_temp[seafloor_temp$reef == 1, ])
+
+  }, FUN.VALUE = numeric(1))
 
   # collapse to charachter string
   no_reefs <- paste(c(no_reefs), collapse = ", ")
 
   # calculate number of fish
-  fish <- lapply(x$fishpop, function(i) {
-
-    # get current fishpop
-    fish_temp <- subset(i, timestep == timestep_temp)
+  fish <- lapply(x$fishpop, function(fish_temp) {
 
     # no individual present, set to 0 and NA
     if (all(is.na(fish_temp[1, -c(18, 19)]))) {
@@ -70,16 +59,21 @@ print.meta_rn <- function(x, digits = 2, timestep = x$max_i, ...) {
         length = round(mean(fish_temp$length), digits = digits),
         mort = round(mean(fish_temp$died_background + fish_temp$died_consumption),
                      digits = digits))
+
     }
   })
 
+  # get minimum timestep
+  min_time <- unique(vapply(X = x$seafloor, function(i) min(i$timestep),
+                            FUN.VALUE = numeric(1)))
+
   # calculate total time
-  total_time <- round(x$max_i * x$min_per_i / 60 / 24, digits = 2)
+  total_time <- round((x$max_i - min_time) * x$min_per_i / 60 / 24, digits = 2)
 
   # calculate save_each timesteps
   save_time <- round(x$save_each * x$min_per_i / 60 / 24, digits = 2)
 
-  cat(paste0("Total time : ", timestep_temp, " iterations (", total_time, " days) [Burn-in: ", x$burn_in, " iter.]\n",
+  cat(paste0("Total time : ", paste0(c(min_time, x$max_i), collapse = "-"), " iterations (", total_time, " days) [Burn-in: ", x$burn_in, " iter.]\n",
              "Saved each : ", x$save_each, " iterations (", save_time, " days)\n",
              "Seafloor   : ", x$extent, "\n",
              "ARs        : ", no_reefs, " cells (movement: ", x$movement, ")\n\n"))